+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Added:
+  * GUI: In the scan results, there is a button to copy registry key paths.
   * CLI: `config path` command to print the path to the active config file.
   * CLI: The `backup` and `restore` commands now have an `--include-disabled` option
     when you want to bulk process games that were disabled in the config.